monet extraction and export works
